+++ NEWS.md
@@ -1,3 +1,8 @@
+# Development version
+
+- Bug-fix for `make_3d_Isingland()`:
+	- The function now allow character values for the `x` and `y`	arguments.
+
 # Isinglandr 0.1.1
 
 - Bug-fix for the Shiny app (`shiny_Isingland_MDD()`):